improve POI processor
